% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscLCA.R
\name{myscLCA}
\alias{myscLCA}
\title{LCA scRNA analysis}
\usage{
myscLCA(datmatrix, clust.max = 10)
}
\arguments{
\item{datmatrix}{data matrix, gene by cell}

\item{clust.max}{maximum number of clusters specified by user}
}
\value{
an integer vector showing clustering membership
}
\description{
This function loads a gene by cell transcript count matarix (row by column).
}
