% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscExampleDat.R
\docType{data}
\name{myscExampleData}
\alias{myscExampleData}
\title{Example single cell RNAseq data}
\format{An object of class \code{list} of length 2.}
\usage{
data(myscExampleData)
}
\description{
The data is a list with a gene-by-cell data.frame and a vector of labels.
The data.frame includes 14,074 transcripts in 250 cells.
While 83% of entries are zero in the gene by cell matrix.
The true labels in three groups for each cell are provided.
}
\examples{
data(myscExampleData)
names(myscExampleData)

}
\keyword{datasets}
